requirements.txt
django-admin
phone_field